Collapse flow to 4 steps: Authenticate & Deploy owns the full CF lifecycle
Resolves the three open design decisions: "Setup access" is removed
(Managed Integrations is the single credentials card, inheriting the
"Steps to assume role" accordion); the Elastic Cloud Forwarder widget on
step 3 now owns launch, the per-service detection animation, and stack
name/version capture; and the old Deploy step is dissolved, with the
Managed Integrations arrival animation merged into Detect & Review's
deployment summary (now step 4). Next on step 3 gates on credentials +
launch + stack name.
